+++ test/karma.conf.js
@@ -83,7 +83,6 @@ module.exports = function(config) {
     // Which plugins to enable
     plugins: [
       'karma-phantomjs-launcher',
-      'karma-chrome-launcher',
       'karma-jasmine',
       'karma-coverage',
       'karma-junit-reporter'